requirements.txt
asgiref==3.7.2
contourpy==1.1.0
cycler==0.11.0
Django==4.2.3
fonttools==4.42.0
importlib-resources==6.0.1
kiwisolver==1.4.4
matplotlib==3.7.2
mysqlclient==2.2.1
numpy==1.25.2
packaging==23.1
Pillow==10.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
zipp==3.16.2
